requirements_current.txt
annotated-types==0.7.0
anyio==4.12.0
blis==1.3.3
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cryptography==46.0.3
cymem==2.0.13
fastapi==0.109.0
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
lxml==6.0.2
MarkupSafe==3.0.3
murmurhash==1.0.15
numpy==2.3.5
packaging==25.0
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==12.0.0
preshed==3.0.12
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pypdfium2==5.2.0
python-docx==1.1.0
python-dotenv==1.2.1
python-multipart==0.0.6
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.4.0
scipy==1.16.3
setuptools==80.9.0
smart_open==7.5.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
starlette==0.35.1
thinc==8.3.10
threadpoolctl==3.6.0
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.27.0
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.3
websockets==15.0.1
wrapt==2.0.1
